*------------------* * Version    1.0.2 * *------------------*  --------------------------------------------  04/04/10 orc@pell.pell.portland.or.us 1.139  configure.sh Check for ndbm.h as part of checking for the ndbm libraries  greylist.c Don't handle nodaemon here  letter.c If nodaemon is set, refuse mail from <> with a 553 (permanent)  letter.c error on the hopes that the client will just shut up and go  letter.c away  --------------------------------------------  04/04/10 orc@pell.pell.portland.or.us 1.140  postoffice.8 Add mailaudit(5) to the crossreference list.  postoffice.8 Don't document -dN, because it's not actually  postoffice.8 used  Makefile.in Install the manpage for the audit record format  audit.c  Change the syslog level to LOG_DEBUG  mailaudit.5 BitKeeper file /home/orc/postoffice/mailaudit.5  --------------------------------------------  04/04/10 orc@downbelow.pell.portland.or.us 1.141  spool.c, recipient.c, mx.c, address.c:    Include "config.h"    On FreeBSD, include <stdlib.h> instead of <malloc.h>  userok.c, newaliases.c, mail.c, letter.c, getif.c:    On FreeBSD, include <stdlib.h> instead of <malloc.h>  configure.sh:    Check for limits.h, and if it's not found try to dummy    up a definition for INT_MAX  mapfd.c:    Pass a size_t pointer instead of a long pointer  env.h:    Pass a size_t pointer in to mapfd  mbox.c:    Include "config.h"    Include <limits.h> instead of <sys/resource.h>    On FreeBSD, include <stdlib.h> instead of <malloc.h>  --------------------------------------------  04/04/10 orc@pell.pell.portland.or.us 1.142  spool.c  Correct size fields so they'll play well with mapfd()  newaliases.c Correct size fields so they'll play well with mapfd()  mail.c   Correct size fields so they'll play well with mapfd()  letter.c Correct size fields so they'll play well with mapfd()  listq.c  Add <string.h> for strchr(), constify the functions  listq.c  passed to scandir to (futilely) stop it from bitching  listq.c  about incompatable types  remote.c Correct size fields so they'll play well with mapfd()  bounce.c Add a prototype for mkaddress()  smtp.c   Correct size fields so they'll play well with mapfd()  letter.h Correct size fields so they'll play well with mapfd()  --------------------------------------------  04/04/11 orc@pell.pell.portland.or.us 1.143  configure.inc Fix configure.inc brokeness for with-foo= when foo  configure.inc includes spaces.  Force a pair of quotes around the  configure.inc field during the eval;  they will get stripped away  configure.inc and leave the value (hopefully) the way it should be  --------------------------------------------  04/04/11 orc@pell.pell.portland.or.us 1.144  Roll up the version number  --------------------------------------------

+++ env.h
@@ -1,6 +1,7 @@
 #ifndef ENV_D
 #define ENV_D
 
+#include <sys/types.h>
 #include <sys/socket.h>
 #include <netinet/in.h>
 #include <arpa/inet.h>
@@ -34,7 +35,7 @@ struct env {
 typedef struct env ENV;
 
 
-char *mapfd(int, long*);
+char *mapfd(int, size_t*);
 
 void set_option(char *, ENV*);
 int  configfile(char *, ENV*);
